+++ R/useful-items.R
@@ -1274,7 +1274,7 @@ attachmentBlock <- function(..., src = NULL, title = NULL, titleUrl = NULL) {
 #'           descriptionBlock(
 #'             number = "17%", 
 #'             numberColor = "success", 
-#'             numberIcon = "fa fa-caret-up",
+#'             numberIcon = "caret-up",
 #'             header = "$35,210.43", 
 #'             text = "TOTAL REVENUE", 
 #'             rightBorder = TRUE,
@@ -1286,7 +1286,7 @@ attachmentBlock <- function(..., src = NULL, title = NULL, titleUrl = NULL) {
 #'           descriptionBlock(
 #'             number = "18%", 
 #'             numberColor = "danger", 
-#'             numberIcon = "fa fa-caret-down",
+#'             numberIcon = "caret-down",
 #'             header = "1200", 
 #'             text = "GOAL COMPLETION", 
 #'             rightBorder = FALSE,
@@ -1319,7 +1319,7 @@ descriptionBlock <- function(number = NULL, numberColor = NULL, numberIcon = NUL
     shiny::tags$span(
       class = numcl, 
       number,
-      if (!is.null(numberIcon)) shiny::tags$i(class = numberIcon)
+      if (!is.null(numberIcon)) shiny::icon(numberIcon)
     ),
     shiny::tags$h5(class = "description-header", header),
     shiny::tags$span(class = "description-text", text)